+++ ErsatzTV.Application/ProgramSchedules/Commands/IProgramScheduleItemRequest.cs
@@ -22,4 +22,5 @@ public interface IProgramScheduleItemRequest
     int? PostRollFillerId { get; }
     int? TailFillerId { get; }
     int? FallbackFillerId { get; }
+    int? WatermarkId { get; }
 }