+++ ErsatzTV.Application/ProgramSchedules/ProgramScheduleItemMultipleViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using ErsatzTV.Application.Filler;
 using ErsatzTV.Application.MediaCollections;
 using ErsatzTV.Application.MediaItems;
+using ErsatzTV.Application.Watermarks;
 using ErsatzTV.Core.Domain;
 
 namespace ErsatzTV.Application.ProgramSchedules;
@@ -25,7 +26,8 @@ public record ProgramScheduleItemMultipleViewModel : ProgramScheduleItemViewMode
         FillerPresetViewModel midRollFiller,
         FillerPresetViewModel postRollFiller,
         FillerPresetViewModel tailFiller,
-        FillerPresetViewModel fallbackFiller) : base(
+        FillerPresetViewModel fallbackFiller,
+        WatermarkViewModel watermark) : base(
         id,
         index,
         startType,
@@ -43,7 +45,8 @@ public record ProgramScheduleItemMultipleViewModel : ProgramScheduleItemViewMode
         midRollFiller,
         postRollFiller,
         tailFiller,
-        fallbackFiller) =>
+        fallbackFiller,
+        watermark) =>
         Count = count;
 
     public int Count { get; }